+++ src/Validator/Constraints/ContentItemMatrixValidator.php
@@ -9,6 +9,8 @@ use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Zicht\Bundle\PageBundle\Entity\ContentItem;
+use Zicht\Bundle\PageBundle\Model\ContentItemContainer;
+use Zicht\Util\Str;
 
 class ContentItemMatrixValidator extends ConstraintValidator
 {
@@ -21,11 +23,10 @@ class ContentItemMatrixValidator extends ConstraintValidator
     }
 
     /**
-     * @param mixed $value
+     * @param ContentItemContainer $value
      */
     public function validate($value, Constraint $constraint)
     {
-        /** @var \Zicht\Bundle\PageBundle\Model\ContentItemMatrix $matrix */
         $matrix = $value->getContentItemMatrix();
 
         if (null !== $matrix) {
@@ -35,12 +36,18 @@ class ContentItemMatrixValidator extends ConstraintValidator
              */
             foreach ($value->getContentItems() as $i => $contentItem) {
                 $type = get_class($contentItem);
+                $typeName = $this->translator->trans('content_item.type.' . strtolower(str_replace(' ', '_', Str::humanize($contentItem->getType()))), [], 'admin');
 
-                if (!in_array($contentItem->getRegion(), $matrix->getRegions($type))) {
-                    $message = $this->translator->trans('content_item.invalid.region_type_combination', ['%region%' => $contentItem->getRegion(), '%type%' => $type], 'validators');
-                    $this->context->buildViolation($message)
-                        ->atPath(sprintf('contentItems[%d]', $i))
-                        ->addViolation();
+                if ((string)$contentItem->getRegion() === '') {
+                    $this->context->buildViolation(
+                        'content_item.invalid.region_empty',
+                        ['%type%' => $typeName]
+                    )->atPath(sprintf('contentItems[%d]', $i))->addViolation();
+                } elseif (!in_array($contentItem->getRegion(), $matrix->getRegions($type))) {
+                    $this->context->buildViolation(
+                        'content_item.invalid.region_type_combination',
+                        ['%region%' => $contentItem->getRegion(), '%type%' => $typeName]
+                    )->atPath(sprintf('contentItems[%d]', $i))->addViolation();
                 }
             }
         }